Move JSON-related JPA tests to their own module
