+++ integration-tests/jpa-postgresql-withjson/pom.xml
@@ -9,8 +9,8 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>quarkus-integration-test-jpa-postgresql</artifactId>
-    <name>Quarkus - Integration Tests - JPA - PostgreSQL</name>
+    <artifactId>quarkus-integration-test-jpa-postgresql-withjson</artifactId>
+    <name>Quarkus - Integration Tests - JPA - PostgreSQL with JSON</name>
     <description>Module that contains JPA related tests running with the PostgreSQL database</description>
 
     <properties>
@@ -29,6 +29,10 @@
         <dependency>
             <groupId>io.quarkus</groupId>
             <artifactId>quarkus-resteasy-reactive</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-jackson</artifactId>
         </dependency>
 
         <!-- test dependencies -->
@@ -83,6 +87,19 @@
         <dependency>
             <groupId>io.quarkus</groupId>
             <artifactId>quarkus-resteasy-reactive-deployment</artifactId>
+            <version>${project.version}</version>
+            <type>pom</type>
+            <scope>test</scope>
+            <exclusions>
+                <exclusion>
+                    <groupId>*</groupId>
+                    <artifactId>*</artifactId>
+                </exclusion>
+            </exclusions>
+        </dependency>
+        <dependency>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-jackson-deployment</artifactId>
             <version>${project.version}</version>
             <type>pom</type>
             <scope>test</scope>
