+++ integration-tests/jpa-postgresql-withjson/src/main/java/io/quarkus/it/jpa/postgresql/otherpu/JsonFormatMapper.java
