+++ integration-tests/jpa-postgresql-withjson/src/main/resources/application.properties
@@ -6,6 +6,9 @@ quarkus.datasource.jdbc.max-size=8
 quarkus.hibernate-orm.packages=io.quarkus.it.jpa.postgresql
 quarkus.hibernate-orm.database.generation=drop-and-create
 quarkus.hibernate-orm.database.generation.create-schemas=true
+# Define non-default PU so that we can configure a custom JSON format mapper. The default PU is using the default mapper.
+quarkus.hibernate-orm."other".datasource=<default>
+quarkus.hibernate-orm."other".packages=io.quarkus.it.jpa.postgresql.otherpu
 
 #Necessary for assertions in JPAFunctionalityInGraalITCase:
 quarkus.native.enable-reports=true
