+++ integration-tests/jpa-postgresql-withjson/src/test/java/io/quarkus/it/jpa/postgresql/JPAFunctionalityInGraalITCase.java
@@ -1,0 +1,11 @@
+package io.quarkus.it.jpa.postgresql;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+/**
+ * Test various JPA operations with JSON running in native mode
+ */
+@QuarkusIntegrationTest
+public class JPAFunctionalityInGraalITCase extends JPAFunctionalityTest {
+
+}
