+++ integration-tests/jpa-postgresql-withjson/src/test/java/io/quarkus/it/jpa/postgresql/JPAFunctionalityTest.java
@@ -1,0 +1,21 @@
+package io.quarkus.it.jpa.postgresql;
+
+import static org.hamcrest.Matchers.is;
+
+import org.junit.jupiter.api.Test;
+
+import io.quarkus.test.junit.QuarkusTest;
+import io.restassured.RestAssured;
+
+/**
+ * Test various JPA operations with JSON running in Quarkus
+ */
+@QuarkusTest
+public class JPAFunctionalityTest {
+
+    @Test
+    public void json() {
+        RestAssured.when().get("/jpa/testfunctionality").then().body(is("OK"));
+    }
+
+}
