+++ integration-tests/jpa-postgresql/src/main/java/io/quarkus/it/jpa/postgresql/JPAFunctionalityTestEndpoint.java
@@ -1,6 +1,5 @@
 package io.quarkus.it.jpa.postgresql;
 
-import java.time.LocalDate;
 import java.util.List;
 import java.util.UUID;
 
@@ -15,8 +14,6 @@ import jakarta.ws.rs.Path;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
 
-import io.quarkus.hibernate.orm.PersistenceUnit;
-import io.quarkus.it.jpa.postgresql.otherpu.EntityWithJsonOtherPU;
 import io.quarkus.narayana.jta.QuarkusTransaction;
 
 /**
@@ -28,10 +25,6 @@ public class JPAFunctionalityTestEndpoint {
 
     @Inject
     EntityManager em;
-
-    @Inject
-    @PersistenceUnit("other")
-    EntityManager otherEm;
 
     @GET
     @Path("base")
@@ -121,51 +114,6 @@ public class JPAFunctionalityTestEndpoint {
                 throw new RuntimeException("Incorrect loaded MyUUIDEntity " + myEntity);
             }
         });
-        return "OK";
-    }
-
-    @GET
-    @Path("json")
-    public String json() {
-        QuarkusTransaction.requiringNew().run(() -> {
-            EntityWithJson entity = new EntityWithJson(
-                    new EntityWithJson.ToBeSerializedWithDateTime(LocalDate.of(2023, 7, 28)));
-            em.persist(entity);
-        });
-
-        QuarkusTransaction.requiringNew().run(() -> {
-            List<EntityWithJson> entities = em
-                    .createQuery("select e from EntityWithJson e", EntityWithJson.class)
-                    .getResultList();
-            if (entities.isEmpty()) {
-                throw new AssertionError("No entities with json were found");
-            }
-        });
-
-        QuarkusTransaction.requiringNew().run(() -> {
-            em.createQuery("delete from EntityWithJson").executeUpdate();
-        });
-
-        Exception exception = null;
-        try {
-            QuarkusTransaction.requiringNew().run(() -> {
-                EntityWithJsonOtherPU otherPU = new EntityWithJsonOtherPU(
-                        new EntityWithJsonOtherPU.ToBeSerializedWithDateTime(LocalDate.of(2023, 7, 28)));
-                otherEm.persist(otherPU);
-            });
-        } catch (Exception e) {
-            exception = e;
-        }
-
-        if (exception == null) {
-            throw new AssertionError(
-                    "Default mapper cannot process date/time properties. So we were expecting transaction to fail, but it did not!");
-        }
-        if (!(exception instanceof UnsupportedOperationException)
-                || !exception.getMessage().contains("I cannot convert anything to JSON")) {
-            throw new AssertionError("flush failed for a different reason than expected.", exception);
-        }
-
         return "OK";
     }
