+++ integration-tests/jpa-postgresql/src/test/java/io/quarkus/it/jpa/postgresql/JPAFunctionalityInGraalITCase.java
@@ -13,7 +13,7 @@ import io.quarkus.test.junit.nativeimage.ClassInclusionReport;
 public class JPAFunctionalityInGraalITCase extends JPAFunctionalityTest {
 
     @Test
-    public void verifyJdkXmlParsersHavebeenEcludedFromNative() {
+    public void verifyJdkXmlParsersHaveBeenExcludedFromNative() {
         final ClassInclusionReport report = ClassInclusionReport.load();
         //The following classes should be included in this applications;
         //if not, that would be a sign that this test has become too weak
