+++ integration-tests/pom.xml
@@ -227,6 +227,7 @@
                 <module>jpa-db2</module>
                 <module>jpa-derby</module>
                 <module>jpa-postgresql</module>
+                <module>jpa-postgresql-withjson</module>
                 <module>jpa-postgresql-withxml</module>
                 <module>jpa-mariadb</module>
                 <module>jpa-h2</module>
